Successfull mirrored projection of our data
Formating our data similar to the wind data, allows us to project our own data.
But the data is mirrored on the x axis, since we have set the latitude starting point at 90.
Also the wind vectors are still there and heavily distorted, this could be because of some unnecessary copy pasting from the original wind data format.